Nodecounts almost match the previous version.

+++ src/move.hpp
@@ -36,13 +36,13 @@ public:
     Square getTo() const;
     Piece getPromotion() const;
     uint16_t getMove() const;
-    int16_t getScore() const;
+    int32_t getScore() const;
     void setMove(uint16_t newMove);
-    void setScore(int16_t newScore);
+    void setScore(int32_t newScore);
     bool empty() const; 
 private:
     uint16_t move;
-    int16_t score;
+    int32_t score;
 };
 
 inline Move::Move() : move(0), score(0)
@@ -52,7 +52,7 @@ inline Move::Move() : move(0), score(0)
 inline Move::Move(const Square from, const Square to, const Piece promotion, const int32_t iScore)
 {
     assert(squareIsOkStrict(from) && squareIsOkStrict(to) && pieceIsOk(promotion));
-    score = static_cast<int16_t>(iScore);
+    score = iScore;
     move = static_cast<uint16_t>(from) | static_cast<uint16_t>(to << 6) | static_cast<uint16_t>(promotion << 12);
 }
 
@@ -76,7 +76,7 @@ inline uint16_t Move::getMove() const
     return move; 
 }
 
-inline int16_t Move::getScore() const 
+inline int32_t Move::getScore() const 
 { 
     return score; 
 }
@@ -86,7 +86,7 @@ inline void Move::setMove(const uint16_t newMove)
     move = newMove; 
 }
 
-inline void Move::setScore(const int16_t newScore) 
+inline void Move::setScore(const int32_t newScore) 
 { 
     score = newScore; 
 }
